requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.1
openai==1.30.1
outcome==1.3.0.post0
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0